requirements.txt
flask==3.1.0
python-dotenv==1.1.0
watchdog==6.0.0
openai==1.77.0